TRMS/trms-back/src/main/java/com/revature/services/UserService.java
package com.revature.services;

import com.revature.beans.Employee;
import com.revature.exceptions.ExistingUsernameException;
import com.revature.exceptions.LoginCredentialsException;

public interface UserService {
	
	public int register(Employee newEmployee) throws ExistingUsernameException;
	public Employee logIn(String username, String password) throws LoginCredentialsException;
	public Employee getUserById(int Id);
	public boolean userUpdate(Employee updateUser);
	public boolean deleteUser(Employee userDelete);
}
